Update snark and fix for macOS
Use https://github.com/zelcash/zelcash-mac to build it.

+++ src/snark/libsnark/algebra/fields/fp12_2over3over2.tcc
@@ -390,7 +390,7 @@ std::istream& operator>>(std::istream& in, std::vector<Fp12_2over3over2_model<n,
 {
     v.clear();
 
-    uint64_t s;
+    size_t s;
     in >> s;
 
     char b;
