+++ src/snark/libsnark/common/utils.cpp
@@ -19,7 +19,7 @@ size_t log2(size_t n)
 /* returns ceil(log2(n)), so UINT64_C(1)<<log2(n) is the smallest power of 2,
    that is not less than n. */
 {
-    uint64_t r = ((n & (n-1)) == 0 ? 0 : 1); // add 1 if n is not power of 2
+    size_t r = ((n & (n-1)) == 0 ? 0 : 1); // add 1 if n is not power of 2
 
     while (n > 1)
     {
@@ -30,10 +30,10 @@ size_t log2(size_t n)
     return r;
 }
 
-uint64_t bitreverse(uint64_t n, const uint64_t l)
+size_t bitreverse(size_t n, const size_t l)
 {
-    uint64_t r = 0;
-    for (uint64_t k = 0; k < l; ++k)
+    size_t r = 0;
+    for (size_t k = 0; k < l; ++k)
     {
         r = (r << 1) | (n & 1);
         n >>= 1;
@@ -44,9 +44,9 @@ uint64_t bitreverse(uint64_t n, const uint64_t l)
 bit_vector int_list_to_bits(const std::initializer_list<uint64_t> &l, const size_t wordsize)
 {
     bit_vector res(wordsize*l.size());
-    for (uint64_t i = 0; i < l.size(); ++i)
+    for (size_t i = 0; i < l.size(); ++i)
     {
-        for (uint64_t j = 0; j < wordsize; ++j)
+        for (size_t j = 0; j < wordsize; ++j)
         {
             res[i*wordsize + j] = (*(l.begin()+i) & (UINT64_C(1)<<(wordsize-1-j)));
         }
@@ -68,7 +68,7 @@ bool is_little_endian()
 
 std::string FORMAT(const std::string &prefix, const char* format, ...)
 {
-    const static uint64_t MAX_FMT = 256;
+    const static size_t MAX_FMT = 256;
     char buf[MAX_FMT];
     va_list args;
     va_start(args, format);
@@ -81,7 +81,7 @@ std::string FORMAT(const std::string &prefix, const char* format, ...)
 void serialize_bit_vector(std::ostream &out, const bit_vector &v)
 {
     out << v.size() << "\n";
-    for (uint64_t i = 0; i < v.size(); ++i)
+    for (size_t i = 0; i < v.size(); ++i)
     {
         out << v[i] << "\n";
     }
@@ -89,10 +89,10 @@ void serialize_bit_vector(std::ostream &out, const bit_vector &v)
 
 void deserialize_bit_vector(std::istream &in, bit_vector &v)
 {
-    uint64_t size;
+    size_t size;
     in >> size;
     v.resize(size);
-    for (uint64_t i = 0; i < size; ++i)
+    for (size_t i = 0; i < size; ++i)
     {
         bool b;
         in >> b;
